All dialog windows moved into appropriate package
