+++ src/ConcreteDialogs/OpenFileDialog.java
@@ -1,3 +1,5 @@
+package ConcreteDialogs;
+
 public class OpenFileDialog extends FileDialog{
     final public String name = "open_file";
